fix error with value types

+++ src/prompt/prompt_engineering.py
@@ -188,9 +188,9 @@ if __name__ == "__main__":
     parser.add_argument("prompt_file_dir", help="directory where json files are stored", type=str)  # directory of prompt json files
     parser.add_argument("prompt_num", help="prompt number", type=int)  # prompt number
     parser.add_argument("llama_cpp", help="llama_cpp", type=int)  # llama_cpp
-    parser.add_argument("top_k", help="top k", type=int)  # top k
-    parser.add_argument("min_p", help="min p", type=int)  # min p
-    parser.add_argument("top_p", help="top p", type=int)  # top p
+    parser.add_argument("top_k", help="top k", type=float)  # top k
+    parser.add_argument("min_p", help="min p", type=float)  # min p
+    parser.add_argument("top_p", help="top p", type=float)  # top p
     parser.add_argument("temperature", help="temperature", type=float)  # temperature
     parser.add_argument('target_names', type=str, help='Comma-separated list of targets') # targets
     parser.add_argument("n_partitions", help="number of partitions", type=int)  # number of partitions
